MNT: re-arranged code and test locations
